+++ src/SubscriptionLink.ts
@@ -1,7 +1,7 @@
 /* eslint-disable class-methods-use-this */
 
 import throttle from 'lodash/throttle'
-import type firebase from 'firebase'
+import type { Firestore } from 'firebase/firestore'
 import { hasDirectives, getOperationName, getOperationDefinition } from 'apollo-utilities'
 import { ApolloLink, Operation, NextLink, Observable, FetchResult } from '@apollo/client'
 
@@ -9,8 +9,8 @@ import parse from './parse'
 import execute from './execute'
 
 export default class SubscriptionLink extends ApolloLink {
-  firestore: firebase.firestore.Firestore
-  constructor({ firestore }: { firestore: firebase.firestore.Firestore }) {
+  firestore: Firestore
+  constructor({ firestore }: { firestore: Firestore }) {
     super()
     this.firestore = firestore
   }