+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### Upcoming
+
+- Upgrade firebase to 9.6.1, requiring peer dependencies to be using version 9^
+
 ### 0.0.4
 
 - Replace `firebase` import with `firebase/app` import